Enhance security and update data types in models
- Updated MongoDB connection string retrieval to use user secrets in `DBContext.cs` and `Program.cs`.
- Added `Orders` DbSet to `DBContext.cs`.
- Changed `number` property type from `double` to `int` in `Products.cs`.
- Removed commented-out database creation line in `Program.cs`.
- Added `UserSecretsId` to the project file for improved configuration management.

+++ WebDev_Labb2_API/Model/Products.cs
@@ -21,7 +21,7 @@ namespace WebDev_Labb2_API.Model
         public double? melt { get; set; }
         public double? molar_heat { get; set; }
         public string? named_by { get; set; }
-        public double number { get; set; }
+        public int number { get; set; }
         public int? period { get; set; }
         public int? group { get; set; }
         public string phase { get; set; }
